requirements.txt
brotlipy==0.7.0
discord.py==1.7.3
multidict==6.0.2
packaging==21.3
python-dotenv==0.20.0
PyYAML==6.0
requests~=2.28.1
numpy~=1.23.1